feat #15
fix double spaces
fix imports with *
fix url of classes
fix implementarion of delete and delete test
fix method "shouldSaveAnUser"
fix the methods on controller test:
     shouldGetAListPage
     shouldListUsers
     saveWithoutName
fix: using just UserLibraryDTO on controller
use of anotattions in atributes to restrict
put private where is the class atribute

+++ src/main/java/com/phoebus/library/userlibrary/service/DeleteUserServiceImpl.java
@@ -1,6 +1,6 @@
 package com.phoebus.library.userlibrary.service;
 
-
+import com.phoebus.library.exceptions.UserNotFoundException;
 import com.phoebus.library.userlibrary.UserLibraryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,6 +13,10 @@ public class DeleteUserServiceImpl implements DeleteUserService{
 
     @Override
     public void delete(Long id) {
-        repository.deleteById(id);
+
+        if (!repository.existsById(id)) {
+            throw new UserNotFoundException();
+        }
+            repository.deleteById(id);
     }
 }
